renamed and abstracted tool logic mte, implemented progress in gui
